refactoring run module
