+++ src/tests/test_run.py
@@ -1,22 +1,25 @@
+import sqlite3
+import warnings
 import logging
 from unittest.mock import Mock
 
+from IPython.core.error import UsageError
 import pandas
 import polars
 import pytest
+from sqlalchemy import create_engine
+import duckdb
 
-import warnings
-
-from IPython.core.error import UsageError
-from sql.connection import Connection
-from sql.run import (
+from sql.connection import Connection, DBAPIConnection
+from sql.run.run import (
     run,
-    handle_postgres_special,
     is_postgres_or_redshift,
     select_df_type,
-    set_autocommit,
+    set_sqlalchemy_autocommit_option,
     display_affected_rowcount,
 )
+from sql.run.pgspecial import handle_postgres_special
+from sql.run.resultset import ResultSet
 
 
 @pytest.fixture
@@ -26,45 +29,38 @@ def mock_conns():
     return conn
 
 
-@pytest.fixture
-def mock_config():
-    class Config:
-        autopandas = None
-        autopolars = None
-        autocommit = True
-        feedback = True
-        polars_dataframe_kwargs = {}
+class Config:
+    autopandas = None
+    autopolars = None
+    autocommit = True
+    feedback = True
+    polars_dataframe_kwargs = {}
+    style = "DEFAULT"
+    autolimit = 0
+    displaylimit = 10
 
-    return Config
+
+class ConfigPandas(Config):
+    autopandas = True
+    autopolars = False
+
+
+class ConfigPolars(Config):
+    autopandas = False
+    autopolars = True
 
 
 @pytest.fixture
 def pytds_conns(mock_conns):
-    mock_conns.dialect = "mssql+pytds"
+    mock_conns._dialect = "mssql+pytds"
     return mock_conns
-
-
-@pytest.fixture
-def config_pandas(mock_config):
-    mock_config.autopandas = True
-    mock_config.autopolars = False
-
-    return mock_config
-
-
-@pytest.fixture
-def config_polars(mock_config):
-    mock_config.autopandas = False
-    mock_config.autopolars = True
-
-    return mock_config
 
 
 @pytest.fixture
 def mock_resultset():
     class ResultSet:
         def __init__(self, *args, **kwargs):
-            ...
+            pass
 
         @classmethod
         def DataFrame(cls):
@@ -73,10 +69,6 @@ def mock_resultset():
         @classmethod
         def PolarsDataFrame(cls):
             return polars.DataFrame()
-
-        @classmethod
-        def fetch_results(self, fetch_all=False):
-            pass
 
     return ResultSet
 
@@ -99,10 +91,10 @@ def test_handle_postgres_special(mock_conns):
     assert "pgspecial not installed" in str(excinfo.value)
 
 
-def test_set_autocommit(mock_conns, mock_config, caplog):
+def test_set_sqlalchemy_autocommit_option(mock_conns, caplog):
     caplog.set_level(logging.DEBUG)
 
-    output = set_autocommit(mock_conns, mock_config)
+    output = set_sqlalchemy_autocommit_option(mock_conns, Config)
 
     with warnings.catch_warnings():
         warnings.simplefilter("error")
@@ -111,9 +103,9 @@ def test_set_autocommit(mock_conns, mock_config, caplog):
     assert output is True
 
 
-def test_pytds_autocommit(pytds_conns, mock_config):
+def test_pytds_autocommit(pytds_conns):
     with warnings.catch_warnings(record=True) as w:
-        output = set_autocommit(pytds_conns, mock_config)
+        output = set_sqlalchemy_autocommit_option(pytds_conns, Config)
         assert (
             str(w[-1].message)
             == "Autocommit is not supported for pytds, thus is automatically disabled"
@@ -121,37 +113,25 @@ def test_pytds_autocommit(pytds_conns, mock_config):
         assert output is False
 
 
-def test_select_df_type_is_pandas(monkeypatch, config_pandas, mock_resultset):
-    monkeypatch.setattr("sql.run.select_df_type", mock_resultset.DataFrame())
-    output = select_df_type(mock_resultset, config_pandas)
+def test_select_df_type_is_pandas(mock_resultset):
+    output = select_df_type(mock_resultset, ConfigPandas)
     assert isinstance(output, pandas.DataFrame)
 
 
-def test_select_df_type_is_polars(monkeypatch, config_polars, mock_resultset):
-    monkeypatch.setattr("sql.run.select_df_type", mock_resultset.PolarsDataFrame())
-    output = select_df_type(mock_resultset, config_polars)
+def test_select_df_type_is_polars(mock_resultset):
+    output = select_df_type(mock_resultset, ConfigPolars)
     assert isinstance(output, polars.DataFrame)
 
 
-def test_sql_starts_with_begin(mock_conns, mock_config):
+def test_sql_starts_with_begin(mock_conns):
     with pytest.raises(UsageError, match="does not support transactions") as excinfo:
-        run(mock_conns, "BEGIN", mock_config)
+        run(mock_conns, "BEGIN", Config)
 
     assert excinfo.value.error_type == "RuntimeError"
 
 
-def test_sql_is_empty(mock_conns, mock_config):
-    assert run(mock_conns, "  ", mock_config) == "Connected: %s" % mock_conns.name
-
-
-def test_run(monkeypatch, mock_conns, mock_resultset, config_pandas):
-    monkeypatch.setattr("sql.run.handle_postgres_special", Mock())
-    monkeypatch.setattr("sql.run._commit", Mock())
-    monkeypatch.setattr("sql.run.display_affected_rowcount", Mock())
-    monkeypatch.setattr("sql.run.ResultSet", mock_resultset)
-
-    output = run(mock_conns, "\\", config_pandas)
-    assert isinstance(output, type(mock_resultset.DataFrame()))
+def test_sql_is_empty(mock_conns):
+    assert run(mock_conns, "  ", Config) == "Connected: %s" % mock_conns.name
 
 
 @pytest.mark.parametrize(
@@ -167,17 +147,35 @@ def test_display_affected_rowcount(capsys, n, message):
     assert captured.out == message
 
 
-def test_commit_is_called(
-    monkeypatch,
-    mock_conns,
-    mock_config,
-):
-    mock__commit = Mock()
-    monkeypatch.setattr("sql.run._commit", mock__commit)
-    monkeypatch.setattr("sql.run.handle_postgres_special", Mock())
-    monkeypatch.setattr("sql.run.display_affected_rowcount", Mock())
-    monkeypatch.setattr("sql.run.ResultSet", Mock())
+@pytest.mark.parametrize(
+    "connection",
+    [
+        Connection(create_engine("duckdb://")),
+        Connection(create_engine("sqlite://")),
+        DBAPIConnection(duckdb.connect()),
+        DBAPIConnection(sqlite3.connect("")),
+    ],
+)
+@pytest.mark.parametrize(
+    "config, expected_type",
+    [
+        [Config, ResultSet],
+        [ConfigPandas, pandas.DataFrame],
+        [ConfigPolars, polars.DataFrame],
+    ],
+)
+@pytest.mark.parametrize(
+    "sql",
+    [
+        "SELECT 1",
+        "SELECT 1; SELECT 2;",
+    ],
+    ids=["single", "multiple"],
+)
+def test_run(connection, config, expected_type, sql):
+    out = run(connection, sql, config)
+    assert isinstance(out, expected_type)
 
-    run(mock_conns, "\\", mock_config)
 
-    mock__commit.assert_called()
+def test_run_calls_commit():
+    pass
